# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GVoice API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
...
  /auth/register:
    post:
      summary: Register account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password]
              properties:
                email: { type: string, format: email }
                username: { type: string, minLength: 3 }
                password: { type: string, minLength: 8 }
      responses:
        '201': { description: Created }
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: OK }
...
  /workspaces/{workspaceId}/channels:
    get:
      summary: List workspace channels
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
    post:
      summary: Create channel
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, type]
              properties:
                name: { type: string }
                type:
                  type: string
                  enum: [text, voice]
                isPrivate: { type: boolean }
      responses:
        '201': { description: Created }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT